# Refactor/optimized model