--- a/mwdb/model/__init__.py
+++ b/mwdb/model/__init__.py
@@ -25,7 +25,7 @@ from .karton import KartonAnalysis, karton_object  # noqa: E402
 from .oauth import OpenIDProvider, OpenIDUserIdentity  # noqa: E402
 from .object import Object, ObjectPermission, relation  # noqa: E402
 from .quick_query import QuickQuery  # noqa: E402
-from .tag import Tag, object_tag_table  # noqa: E402
+from .tag import Tag  # noqa: E402
 from .user import User  # noqa: E402
 
 __all__ = [
@@ -50,6 +50,5 @@ __all__ = [
     "relation",
     "QuickQuery",
     "Tag",
-    "object_tag_table",
     "User",
 ]