--- a/mwdb/model/file.py
+++ b/mwdb/model/file.py
@@ -31,25 +31,21 @@ class EmptyFileError(ValueError):
 
 
 class File(Object):
-    __tablename__ = "file"
-
-    id = db.Column(db.Integer, db.ForeignKey("object.id"), primary_key=True)
-    file_name = db.Column(db.String, nullable=False, index=True)
-    file_size = db.Column(db.Integer, nullable=False, index=True)
-    file_type = db.Column(db.Text, nullable=False, index=True)
-    md5 = db.Column(db.String(32), nullable=False, index=True)
-    crc32 = db.Column(db.String(8), nullable=False, index=True)
-    sha1 = db.Column(db.String(40), nullable=False, index=True)
-    sha256 = db.Column(db.String(64), nullable=False, index=True, unique=True)
-    sha512 = db.Column(db.String(128), nullable=False, index=True)
-    # ssdeep is nullable due to lack of support in earlier versions
-    ssdeep = db.Column(db.String(255), nullable=True, index=True)
+    file_name = db.Column(db.String, index=True)
+    file_size = db.Column(db.Integer, index=True)
+    file_type = db.Column(db.Text, index=True)
+    md5 = db.Column(db.String(32), index=True)
+    crc32 = db.Column(db.String(8), index=True)
+    sha1 = db.Column(db.String(40), index=True)
+    sha256 = db.Column(db.String(64), index=True, unique=True)
+    sha512 = db.Column(db.String(128), index=True)
+    ssdeep = db.Column(db.String(255), index=True)
     alt_names = db.Column(
         MutableList.as_mutable(ARRAY(db.String)), nullable=False, server_default="{}"
     )
 
     __mapper_args__ = {
-        "polymorphic_identity": __tablename__,
+        "polymorphic_identity": "file",
     }
 
     @classmethod