<!-- Thank you for contributing! -->
<!-- Please fill this template before submitting your PR (fill the boxes using "x") -->

**Your checklist for this pull request**
- [x] I've read the [contributing guideline](CONTRIBUTING.md).
- [x] I've tested my changes by building and running the project, and testing changed functionality (if applicable)
- [ ] I've added automated tests for my change (if applicable, optional)
- [ ] I've updated documentation to reflect my change (if applicable)

**What is changed?**

Tag queries are slow as described in #654. In this PR I try to optimize model to avoid heavy joins that may occur during queries.

Changes in model:

- Object<->Tag relation was converted from many-to-many to one-to-many. Tag is represented by (object_id, tag_string) association instead of (object_id, tag_id) with tag in separate Table.
- Converted Object<->File/Config/Blob inheritance model from join-based to single-table based. All information is contained in single table Object instead of using separate tables for specialized fields, joined with common primary key.

After doing basic benchmarks on production-like data and I noticed serious performance improvement:
```
Before optimization ===================

Querying for '': count 2511234 took 1.4322917461395264s, query took 0.015579938888549805s
Querying for 'tag:"feed:malwarebazaar"': count 512790 took 2.299734115600586s, query took 0.01782536506652832s
Querying for 'tag:"feed:malwarebazaar" AND upload_time:<=2022-06-01': count 466728 took 2.5679376125335693s, query took 0.07134366035461426s
Querying for 'tag:"feed:*" AND tag:"*exe"': count 328541 took 2.3470044136047363s, query took 0.1802363395690918s
Querying for 'parent:(tag:"feed:*" AND tag:"*exe")': count 110701 took 3.2340664863586426s, query took 3.855536460876465s

After optimization ====================
Querying for '': count 2511234 took 1.162876844406128s, query took 0.015012502670288086s
Querying for 'tag:"feed:malwarebazaar"': count 512790 took 1.4038827419281006s, query took 0.02241373062133789s
Querying for 'tag:"feed:malwarebazaar" AND upload_time:<=2022-06-01': count 466728 took 1.205815076828003s, query took 0.1941375732421875s
Querying for 'tag:"feed:*" AND tag:"*exe"': count 328541 took 1.4597358703613281s, query took 0.017667055130004883s
Querying for 'parent:(tag:"feed:*" AND tag:"*exe")': count 110701 took 3.6421310901641846s, query took 0.049126625061035156s
```

Benchmark was done using the following script:
```python
import time
from mwdblib import MWDB

queries = [
    '',
    'tag:"feed:malwarebazaar"',
    'tag:"feed:malwarebazaar" AND upload_time:<=2022-06-01',
    'tag:"feed:*" AND tag:"*exe"',
    'parent:(tag:"feed:*" AND tag:"*exe")'
]

mwdb = MWDB(api_url="http://127.0.0.1/api/")
# mwdb.login()

for q in queries:
    print(f"Querying for '{q}': ", end='', flush=True)
    start = time.time()
    count = mwdb.count_files(q)
    stop = time.time()
    print(f"count {count} took {stop - start}s, ", end='', flush=True)
    start = time.time()
    next(mwdb.search_files(q))
    stop = time.time()
    print(f"query took {stop - start}s")
```

During migration, database size changed from 8.5GB to 9.7GB (with 14GB peak during migration), but after VACUUM FULL was reduced to 8.7GB. It's worth noticing that we need a double of space for migration because major data must be copied from one table to another.

TODO:
- [ ] Migrations has no downgrade scripts, maybe we should consider to write some
- [ ] Documentation notes

**Closing issues**

<!-- Add in issue numbers related to this PR, if applicable -->

closes #654